Vue: Fold the render-less story path into the template engine
